Prevent an error in the call to (DEFCONSTANT +ERROR-RETURN+)
The initialisation form to DEFCONSTANT calls ICONV-OPEN. Since the last fix,
this function is now a normal function instead of a CFFI function. This meant
that this function was not available while the DEFCONSTANT initialisation form
was evaluated.

This fix places this function inside an EVAL-WHEN construct to make sure it's
available to the subsequent call to DEFCONSTANT.