Make minimum size of the output buffer 100 bytes
Under some circumstances iconv can fail if the buffer is too small,
even if it's big enough to contain the entire output data. Specifically,
this happens if the output data is one byte, and the output buffer is
two bytes in size. This would read to an infinite loop where CL-ICONV
would keep trying to call iconv with the same arguments.

This fix makes sure the output buffer is never smaller than 100 bytes,
which will prevent this problem from happening.